platform/backend/pyproject.toml
[project]
name = "job-platform"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic-settings>=2.6",
    "authlib>=1.4",
    "httpx>=0.28",
    "python-jose[cryptography]>=3.3",
    "celery[redis]>=5.4",
    "redis>=5.2",
    "pyyaml>=6.0",
    "openpyxl>=3.1",
    "python-multipart>=0.0.12",
    "beautifulsoup4>=4.12",
    "dnspython>=2.6",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "httpx"]
